refactor(mobile): move 2fa as modal
